Add files via upload
Version 1.1 Description - A simple stats program for crunching the min, max and mean of a variable set of numbers provided by the user. Version 1.1 is designed for a singular execution. Will expand in the future with a looping feature probably until user exits. This version added mode and median to the mix.